File rotation
This patch adds a new "periodic" manager, which automatically takes a
snapshot of all of the measurements in a context at some pre-defined
interval.  That logic can be executed in a background thread, or you can
trigger interval checks manually for test cases or to integrate with an
existing event loop.

Building on that, there's also a new "file rotator", which writes the
value of each measurement in the snapshot to a file in some output
directory, rotating those files after a certain duration.  The file
rotator uses a periodic manager under the covers, so it gets the
interval checks for free.

End result is that you can spin up a background thread that writes the
current value of each measurement into a file every (for instance) 10
seconds, while only keeping any given file open for (for instance) 5
minutes.