Final snapshot in periodic background thread
The periodic manager's background thread will now always fire its
callback one final time when the thread is canceled.  For the file
rotator, this means that the last file will have a final entry
containing the values of the statistics when the background thread was
canceled (which is usually at the end of the process).