Rendering deltas with rates
Given two snapshots, and the interval between them, we can compute the
rate that derive counters are increasing by.  This patch adds new
functions for rendering two snapshots that includes rates for any
derives.  (The rate portion is left blank for gauges.)

The periodic interface has also been updated to keep track of the
previous snapshot, along with the current one, to make it easy for
background threads to calculate rates.