Provide a filename slug for stat files
We provide the ability to add slug values to fathom-sensor data files,
so we should do the same for our stats files.  It makes it much easier
to track and correlate them across customes and time.

This patch provides a setter method for a filename slug.  The slug value
is placed between the "stats" and timestamp substrings in the filename.